pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-dev-agent"
version = "0.1.0"
description = "Python CLI agent for software development automation"
authors = [{name = "Codex"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "click>=8.1",
  "requests>=2.31",
  "pydantic>=2.8",
  "tomli; python_version < '3.11'",
  "tree-sitter==0.20.1",
  "tree-sitter-languages>=1.10",
  "pyyaml>=6.0"
]

[project.scripts]
devagent = "ai_dev_agent.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_dev_agent*"]

[project.optional-dependencies]
dev = [
  "pytest>=7.4"
]
